data/requirements.txt
algoliasearch==2.1.0
apache-airflow[crypto,postgres,password]==1.10.5
google-cloud-bigquery==1.20.0
